Added global exception handlers and enhanced data parser exception handling.
- Implemented global exception handlers for HTTP, validation, and general errors in FastAPI.
- Added specific exception handling in the data parser to manage errors when retrieving weather data, including custom error messages with city names.
- Added an error page (`error.html`) to display error messages to users in case of failed weather data retrieval.

+++ models/weather.py
@@ -6,10 +6,10 @@ from datetime import datetime
 # Define the CurrentWeather model to represent the weather data
 class CurrentWeather(BaseModel):
     city_name: str
-    temperature: float
-    feels_like: float  # Temperature feels like
-    humidity: float
-    wind: float
+    temperature: int
+    feels_like: int  # Temperature feels like
+    humidity: int
+    wind: int
     description: str
     icon_url: str
 
@@ -17,7 +17,7 @@ class CurrentWeather(BaseModel):
 # Define the model for a single forecast entry
 class ForecastEntry(BaseModel):
     date_time: datetime
-    temperature: float
+    temperature: int
     description: str
     icon_url: str
